+++ proofs/files.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 use Innmind\IO\{
     IO,
     Files\Read,
+    Files\Temporary,
     Files\Temporary\Pull,
 };
 use Innmind\Url\Path;
@@ -347,6 +348,45 @@ return static function() {
             $assert->same(
                 $expected,
                 $read,
+            );
+        },
+    );
+
+    yield proof(
+        'IO::files()->temporary()->close()',
+        given($strings),
+        static function($assert, $chunks) {
+            $tmp = IO::fromAmbientAuthority()
+                ->files()
+                ->temporary(Sequence::of(...$chunks)->map(Str::of(...)))
+                ->match(
+                    static fn($tmp) => $tmp,
+                    static fn() => null,
+                );
+
+            $assert
+                ->object($tmp)
+                ->instance(Temporary::class);
+
+            $assert->not()->null(
+                $tmp->read()->size()->match(
+                    static fn($size) => $size,
+                    static fn() => null,
+                ),
+            );
+
+            $assert
+                ->object($tmp->close()->match(
+                    static fn($sideEffect) => $sideEffect,
+                    static fn() => null,
+                ))
+                ->instance(SideEffect::class);
+
+            $assert->null(
+                $tmp->read()->size()->match(
+                    static fn($size) => $size,
+                    static fn() => null,
+                ),
             );
         },
     );