+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\IO\Files\Temporary::close()`
+
 ## 3.1.0 - 2025-04-21
 
 ### Added